add lsh

+++ ImageSearch/src/training/util.h
@@ -9,6 +9,6 @@
 
 #define assertion(test) do{if(!(test)){outside_util::printf("Assertion failed!");outside_util::exit(1);}}while(0)
 
-float calc_distance(const float* a, const float* b, size_t n);
+double calc_distance(const float* a, const float* b, size_t n);
 
 #endif
